modularize redux code
