+++ frontend/src/app/app.jsx
@@ -1,49 +1,9 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { createStore } from 'redux';
+import reducer from './reducers/reducer';
+import { actionCreators } from './actions/index';
 import WodBox from './components/wodBox';
-
-const actionTypes = {
-  SELECT_RANDOM_WOD: 'SELECT_RANDOM_WOD',
-  SET_WODS: 'SET_WODS'
-};
-
-const actionCreators = {
-  selectRandomWod: () => ({
-    type: actionTypes.SELECT_RANDOM_WOD
-  }),
-  setWods: (wods) => ({
-    type: actionTypes.SET_WODS,
-    wods
-  })
-};
-
-const defaultState = {
-  randomWod: null,
-  wods: []
-};
-
-const reducer = (state = defaultState, action) => {
-  const {
-    randomWod,
-    type,
-    wods
-  } = action;
-
-  switch (type) {
-    case actionTypes.SELECT_RANDOM_WOD:
-      return Object.assign({}, state, {
-        randomWod: state.wods[Math.floor(Math.random() * state.wods.length)].name
-    });
-
-    case actionTypes.SET_WODS:
-      return Object.assign({}, state, {
-        wods
-    });
-  }
-
-  return state;
-};
 
 const store = createStore(reducer);
