+++ webpack-stats.json
@@ -1,1 +1,1 @@
-{"status":"done","chunks":{"main":[{"name":"main-51628f497e3d46b2c1ad.js","path":"/Users/angelika/wod/frontend/src/bundles/main-51628f497e3d46b2c1ad.js"}]}}
+{"status":"done","chunks":{"main":[{"name":"main-0bb6f83ba69d8e844460.js","path":"/Users/angelika/wod/frontend/src/bundles/main-0bb6f83ba69d8e844460.js"}]}}
